+++ src/main/java/verbly/spring/domain/home/controller/HomeController.java
@@ -1,17 +1,13 @@
-package verbly.spring.domain.user.controller;
+package verbly.spring.domain.home.controller;
 
 import lombok.RequiredArgsConstructor;
-import lombok.Value;
-import org.springframework.data.repository.query.Param;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import verbly.spring.domain.home.dto.response.HomeResponseDTO;
 import verbly.spring.domain.stats.service.StatsCommandService;
-import verbly.spring.domain.user.dto.response.UserResponseDTO;
-import verbly.spring.domain.user.service.userhome.UserHomeQueryService;
+import verbly.spring.domain.home.service.UserHomeQueryService;
 import verbly.spring.global.common.response.ApiResponse;
-import verbly.spring.global.security.auth.CustomUserDetails;
 import verbly.spring.global.security.utils.SecurityUtils;
 
 import java.util.UUID;
@@ -20,21 +16,21 @@ import java.util.UUID;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/home")
-public class UserHomeController implements UserControllerDocs {
+public class HomeController implements HomeControllerDocs {
 
     private final UserHomeQueryService userHomeQueryService;
     private final StatsCommandService statsCommandService;
 
     @Override
     @GetMapping("/viewer/info")
-    public ApiResponse<UserResponseDTO.HomeViewerInfoDTO> getHomeViewerInfo(
+    public ApiResponse<HomeResponseDTO.HomeViewerInfoDTO> getHomeViewerInfo(
     ){
         return ApiResponse.onSuccess(userHomeQueryService.getHomeViewerInfo());
     }
 
     @Override
     @GetMapping("/users/{uuid}")
-    public ApiResponse<UserResponseDTO.HomeUserInfoDTO> getUserProfileInfo(
+    public ApiResponse<HomeResponseDTO.HomeUserInfoDTO> getUserProfileInfo(
             @PathVariable(name = "uuid") UUID uuid
     ){
         return ApiResponse.onSuccess(userHomeQueryService.getUserProfileInfo(uuid));
