+++ src/main/java/verbly/spring/domain/user/converter/UserConverter.java
@@ -72,32 +72,4 @@ public class UserConverter {
                 )
                 .build();
     }
-
-    public static UserResponseDTO.HomeViewerInfoDTO toHomeViewerInfoDTO(User viewer, long following, long correctionReceived) {
-        return UserResponseDTO.HomeViewerInfoDTO.builder()
-                .imageUrl(viewer.getProfileImage().getImageUrl())
-                .nickname(viewer.getNickname())
-                .following((int)following)
-                .streak(viewer.getStats().getStreakDays())
-                .point(viewer.getStats().getPoint())
-                .correctionReceived((int)correctionReceived)
-                .level(viewer.getStats().getLevel())
-                .build();
-    }
-
-    public static UserResponseDTO.HomeUserInfoDTO toHomeUserInfoDTO(User target, long totalPosts, long following, long follower,
-                                                                    boolean isFollowing, long correctionReceived, long correctionGiven) {
-        return UserResponseDTO.HomeUserInfoDTO.builder()
-                .imageUrl(target.getProfileImage().getImageUrl())
-                .nickname(target.getNickname())
-                .nativeLang(target.getNativeLang())
-                .description(target.getBio())
-                .totalPosts((int)totalPosts)
-                .follower((int)follower)
-                .following((int)following)
-                .isFollowing(isFollowing)
-                .correctionReceived((int)correctionReceived)
-                .correctionGiven((int)correctionGiven)
-                .build();
-    }
 }
